Refactor pool-specific data into a single module
Deployments should use the pool-info module to pick up the pool ID,
primary shared VM disk storage SR ID and external/public network ID.
These vary between pools, naturally, and the values required to look
them up can also vary.

The SRCF only has three distinct pools, so it's feasible to embed the
knowledge required to look up the relevant info for each pool by its
name into a single module, which is the function pool-info performs.

If there are new parts of generic infrastructure available in SRCF VM
pools, then this module can be augmented to publish information on them.
This may be, for example, a "private network ID" (which could be null to
signify that no such infrastructure is available in a given pool).

This change does not alter the Terraform plan; in other words it does
not require a redeployment.